renamed to undefined and releasing to npme.one.com
As we are changing to npm Enterprise, we have to rename all packages and
republish them to the new registry. This was done as an automated effort.

Older versions of this package has been published to the registry and rewritten
the same way as this commit. To reduce noise it is only landed on the tip of
master and not on old feature branches.